Restructure datastore code and add AutoStore
