--- src/datastore/cache/clear.js
@@ -1,5 +1,0 @@
-export async function clear() {
-  const { appKey } = getConfig();
-  await _clear(appKey);
-  return null;
-}
