+++ src/datastore/index.js
@@ -1,7 +1,7 @@
 import isString from 'lodash/isString';
 import KinveyError from '../errors/kinvey';
-import { isValidTag, clear as _clear } from './cache';
-import { NetworkStore } from './networkstore';
+import { clear as _clear } from './cache';
+import NetworkStore from './networkstore';
 import { SyncStore } from './syncstore';
 import AutoStore from './autostore';
 
@@ -13,22 +13,14 @@ export const DataStoreType = {
 
 export function collection(collectionName, type = DataStoreType.Cache, options = {}) {
   let datastore;
-  const tagWasPassed = options && ('tag' in options);
 
-  if (collectionName == null || !isString(collectionName)) {
+  if (!collectionName || !isString(collectionName)) {
     throw new KinveyError('A collection is required and must be a string.');
-  }
-  if (tagWasPassed && !isValidTag(options.tag)) {
-    throw new KinveyError('Please provide a valid data store tag.');
   }
 
   if (type === DataStoreType.Auto) {
     datastore = new AutoStore(collectionName, Object.assign({}, options));
   } else if (type === DataStoreType.Network) {
-    if (tagWasPassed) {
-      throw new KinveyError('The tagged option is not valid for data stores of type "Network"');
-    }
-
     datastore = new NetworkStore(collectionName);
   } else if (type === DataStoreType.Sync) {
     datastore = new SyncStore(collectionName, Object.assign({}, options));
@@ -37,10 +29,6 @@ export function collection(collectionName, type = DataStoreType.Cache, options =
   }
 
   return datastore;
-}
-
-export function getInstance(collection, type, options) {
-  return collection(collection, type, options);
 }
 
 export async function clear() {
