--- src/datastore/network/index.js
@@ -1,19 +1,0 @@
-import count from './count';
-import create from './create';
-import find from './find';
-import findById from './findById';
-import group from './group';
-import remove from './remove';
-import removeById from './removeById';
-import update from './update';
-
-export {
-  count,
-  create,
-  find,
-  findById,
-  group,
-  remove,
-  removeById,
-  update
-};
