+++ src/datastore/sync.js
@@ -1,9 +1,19 @@
-import Query from '../query';
+import times from 'lodash/times';
+import isArray from 'lodash/isArray';
 import SyncError from '../errors/sync';
+import KinveyError from '../errors/kinvey';
+import MissingConfigurationError from '../errors/missingConfiguration';
+import ParameterValueOutOfRangeError from '../errors/parameterValueOutOfRange';
 import NotFoundError from '../errors/notFound';
-import { NetworkStore } from './networkstore';
-import { DataStoreCache } from './cache';
+import Query from '../query';
+import { KinveyHeaders } from '../http/headers';
+import SyncCache from './syncCache';
+import SyncEvent from './syncEvent';
+import Network from './network';
+import Cache from './cache';
+import QueryCache from './queryCache';
 
+const PAGE_LIMIT = 10000;
 const PUSH_IN_PROGRESS = {};
 
 function markPushStart(collectionName) {
@@ -14,36 +24,30 @@ function markPushEnd(collectionName) {
   PUSH_IN_PROGRESS[collectionName] = false;
 }
 
-export function queryToSyncQuery(query, collectionName) {
-  if (query && query instanceof Query) {
-    const newFilter = Object.keys(query.filter)
-      .reduce((filter, field) => Object.assign({}, filter, { [`entity.${field}`]: query.filter[field] }), {});
-    const newSort = Object.keys(query.sort)
-      .reduce((sort, field) => Object.assign({}, sort, { [`entity.${field}`]: query.sort[field] }), {});
-    const syncQuery = new Query({
-      filter: newFilter,
-      sort: newSort,
-      skip: query.skip,
-      limit: query.limit
-    });
+// export function queryToSyncQuery(query, collectionName) {
+//   if (query && query instanceof Query) {
+//     const newFilter = Object.keys(query.filter)
+//       .reduce((filter, field) => Object.assign({}, filter, { [`entity.${field}`]: query.filter[field] }), {});
+//     const newSort = Object.keys(query.sort)
+//       .reduce((sort, field) => Object.assign({}, sort, { [`entity.${field}`]: query.sort[field] }), {});
+//     const syncQuery = new Query({
+//       filter: newFilter,
+//       sort: newSort,
+//       skip: query.skip,
+//       limit: query.limit
+//     });
 
-    if (collectionName) {
-      syncQuery.equalTo('collection', collectionName);
-    }
+//     if (collectionName) {
+//       syncQuery.equalTo('collection', collectionName);
+//     }
 
-    return syncQuery;
-  }
+//     return syncQuery;
+//   }
 
-  return null;
-}
+//   return null;
+// }
 
-export const SyncEvent = {
-  Create: 'POST',
-  Update: 'PUT',
-  Delete: 'DELETE'
-};
-
-export class Sync {
+export default class Sync {
   constructor(collectionName, tag) {
     this.collectionName = collectionName;
     this.tag = tag;
@@ -86,7 +90,7 @@ export class Sync {
     let syncDocs = [];
     let docsToSync = docs;
 
-    if (!Array.isArray(docs)) {
+    if (!isArray(docs)) {
       singular = true;
       docsToSync = [docs];
     }
@@ -129,8 +133,8 @@ export class Sync {
   }
 
   async push(query, options) {
-    const network = new NetworkStore(this.collectionName);
-    const cache = new DataStoreCache(this.collectionName, this.tag);
+    const network = new Network(this.collectionName);
+    const cache = new Cache(this.collectionName, this.tag);
     const syncCache = new SyncCache(this.tag);
 
     if (this.isPushInProgress()) {
@@ -252,6 +256,176 @@ export class Sync {
     }
 
     return [];
+  }
+
+  async pull(query, options = {}) {
+    const network = new Network(this.collectionName);
+    const cache = new Cache(this.collectionName, this.tag);
+    const queryCache = new QueryCache(this.tag);
+    const useDeltaSet = options.useDeltaSet === true || this.useDeltaSet;
+    const useAutoPagination = options.useAutoPagination === true || options.autoPagination || this.useAutoPagination;
+
+    // Push sync queue
+    const count = await this.pendingSyncCount();
+    if (count > 0) {
+      // TODO in newer version
+      // if (autoSync) {
+      //   await sync.push();
+      //   return this.pull(query, Object.assign({}, { useDeltaSet, useAutoPagination, autoSync }, options));
+      // }
+
+      if (count === 1) {
+        throw new KinveyError(`Unable to pull entities from the backend. There is ${count} entity`
+          + ' that needs to be pushed to the backend.');
+      }
+
+      throw new KinveyError(`Unable to pull entities from the backend. There are ${count} entities`
+        + ' that need to be pushed to the backend.');
+    }
+
+    // Delta set
+    if (useDeltaSet && (!query || (query.skip === 0 && query.limit === Infinity))) {
+      try {
+        const key = queryCache.serializeQuery(query);
+        const queryCacheDoc = await queryCache.findByKey(key);
+
+        if (queryCacheDoc && queryCacheDoc.lastRequest) {
+          const response = await network.findWithDeltaSet(query, Object.assign({}, options, { since: queryCacheDoc.lastRequest }));
+          const { changed, deleted } = response.data;
+
+          // Delete the docs that have been deleted
+          if (Array.isArray(deleted) && deleted.length > 0) {
+            const removeQuery = new Query().contains('_id', deleted.map(doc => doc._id));
+            await cache.remove(removeQuery);
+          }
+
+          // Save the docs that changed
+          if (Array.isArray(changed) && changed.length > 0) {
+            await cache.save(changed);
+          }
+
+          // Update the query cache
+          const headers = new KinveyHeaders(response.headers);
+          queryCacheDoc.lastRequest = headers.requestStart;
+          await queryCache.save(queryCacheDoc);
+
+          // Return the number of changed docs
+          return changed.length;
+        }
+      } catch (error) {
+        if (!(error instanceof MissingConfigurationError) && !(error instanceof ParameterValueOutOfRangeError)) {
+          throw error;
+        }
+      }
+    }
+
+    // Auto pagination
+    if (useAutoPagination) {
+      // Clear the cache
+      await cache.clear();
+
+      // Get the total count of docs
+      const response = await network.count(query, Object.assign({}, options, { rawResponse: true }));
+      const count = 'count' in response.data ? response.data.count : Infinity;
+
+      // Create the pages
+      const pageSize = options.autoPaginationPageSize || (options.autoPagination && options.autoPagination.pageSize) || PAGE_LIMIT;
+      const pageCount = Math.ceil(count / pageSize);
+      const pageQueries = times(pageCount, (i) => {
+        const pageQuery = new Query(query);
+        pageQuery.skip = i * pageSize;
+        pageQuery.limit = Math.min(count - (i * pageSize), pageSize);
+        return pageQuery;
+      });
+
+      // Process the pages
+      const pagePromises = pageQueries.map((pageQuery) => {
+        return network.find(pageQuery, options)
+          .then(docs => cache.save(docs))
+          .then(docs => docs.length);
+      });
+      const pageCounts = await Promise.all(pagePromises);
+      const totalPageCount = pageCounts.reduce((totalCount, pageCount) => totalCount + pageCount, 0);
+
+      // Update the query cache
+      const key = queryCache.serializeQuery(query);
+      let queryCacheDoc = await queryCache.findByKey(key);
+      if (!queryCacheDoc) {
+        queryCacheDoc = { collectionName: this.collectionName, query: key };
+      }
+      const headers = new KinveyHeaders(response.headers);
+      queryCacheDoc.lastRequest = headers.requestStart;
+      await queryCache.save(queryCacheDoc);
+
+      // Return the total page count
+      return totalPageCount;
+    }
+
+    // Find the docs on the backend
+    const response = await network.find(query, Object.assign({}, options, { rawResponse: true }));
+    const docs = response.data;
+
+    // Clear the cache if a query was not provided
+    if (!query) {
+      await cache.clear();
+    }
+
+    // Update the cache
+    await cache.save(docs);
+
+    // Update the query cache
+    const key = queryCache.serializeQuery(query);
+    let queryCacheDoc = await queryCache.findByKey(key);
+    if (!queryCacheDoc) {
+      queryCacheDoc = { collectionName: this.collectionName, query: key };
+    }
+    const headers = new KinveyHeaders(response.headers);
+    queryCacheDoc.lastRequest = headers.requestStart;
+    await queryCache.save(queryCacheDoc);
+
+    // Return the number of docs
+    return docs.length;
+  }
+
+  async pullById(id, options = {}) {
+    const network = new Network(this.collectionName);
+    const cache = new Cache(this.collectionName, this.tag);
+
+    // Push sync queue
+    const count = await this.pendingSyncCount();
+    if (count > 0) {
+      // TODO in newer version
+      // if (autoSync) {
+      //   await sync.push();
+      //   return this.pull(query, Object.assign({}, { useDeltaSet, useAutoPagination, autoSync }, options));
+      // }
+
+      if (count === 1) {
+        throw new KinveyError(`Unable to pull entities from the backend. There is ${count} entity`
+          + ' that needs to be pushed to the backend.');
+      }
+
+      throw new KinveyError(`Unable to pull entities from the backend. There are ${count} entities`
+        + ' that need to be pushed to the backend.');
+    }
+
+    try {
+      // Find the doc on the backend
+      const doc = await network.findById(id, options);
+
+      // Update the doc in the cache
+      await cache.save(doc);
+
+      // Return the doc
+      return doc;
+    } catch (error) {
+      if (error instanceof NotFoundError) {
+        // Remove the doc from the cache
+        await cache.removeById(id);
+      }
+
+      throw error;
+    }
   }
 
   async remove(query) {
