--- src/datastore/sync/addCreateSyncEvent.js
@@ -1,3 +1,0 @@
-export default function addCreateSyncEvent(docs) {
-  return addSyncEvent(SyncEvent.Create, docs);
-}
