--- src/datastore/sync/count.js
@@ -1,6 +1,0 @@
-import SyncCache from './syncCache';
-
-export default function count(tag, query) {
-  const syncCache = new SyncCache(tag);
-  return syncCache.count(query);
-}
