--- src/datastore/sync/find.js
@@ -1,6 +1,0 @@
-import SyncCache from './syncCache';
-
-export default function find(tag, query) {
-  const syncCache = new SyncCache(tag);
-  return syncCache.find(query);
-}
