--- src/datastore/sync/findById.js
@@ -1,6 +1,0 @@
-import SyncCache from './syncCache';
-
-export default function findById(tag, id) {
-  const syncCache = new SyncCache(tag);
-  return syncCache.findById(id);
-}
