+++ src/datastore/syncEvent.js
@@ -1,0 +1,6 @@
+const SyncEvent = {
+  Create: 'POST',
+  Update: 'PUT',
+  Delete: 'DELETE'
+};
+export default SyncEvent;
